Updating `ex`.
