--- ex/COVERAGE
@@ -1,1 +1,0 @@
-72.2
