+++ ex/class.go
@@ -2,11 +2,6 @@ package ex
 
 import "encoding/json"
 
-// ClassProvider is a type that can return an exception class.
-type ClassProvider interface {
-	Class() error
-}
-
 // Class is a string wrapper that implements `error`.
 // Use this to implement constant exception causes.
 type Class string
