+++ ex/class_provider.go
@@ -1,0 +1,6 @@
+package ex
+
+// ClassProvider is a type that can return an exception class.
+type ClassProvider interface {
+	Class() error
+}
