+++ ex/err_message.go
@@ -1,0 +1,14 @@
+package ex
+
+// ErrMessage returns the exception message.
+// This depends on if the err is itself an exception or not.
+// If it is not an exception, this will return empty string.
+func ErrMessage(err interface{}) string {
+	if err == nil {
+		return ""
+	}
+	if ex := As(err); ex != nil && ex.Class != nil {
+		return ex.Message
+	}
+	return ""
+}
