+++ ex/inner_provider.go
@@ -1,0 +1,6 @@
+package ex
+
+// InnerProvider is a type that returns an inner error.
+type InnerProvider interface {
+	Inner() error
+}
