+++ ex/main_test.go
@@ -1,0 +1,11 @@
+package ex
+
+import (
+	"os"
+	"testing"
+)
+
+// TestMain is the testing entrypoint.
+func TestMain(m *testing.M) {
+	os.Exit(m.Run())
+}
