--- ex/package.go
@@ -1,3 +1,0 @@
-// Package ex adds the ability to wrap errors with stack traces, structured error classes and free form descriptive messages.
-// Most packages in go-sdk return exceptions instead of strict string errors.
-package ex
