+++ ex/util_test.go
@@ -7,17 +7,6 @@ import (
 	"github.com/blend/go-sdk/assert"
 )
 
-func TestIs(t *testing.T) {
-	assert := assert.New(t)
-
-	errInvalidSomething := Class("invalid something")
-
-	ex := New(errInvalidSomething)
-
-	assert.True(Is(ex, errInvalidSomething))
-	assert.True(Is(errInvalidSomething, errInvalidSomething))
-}
-
 type classProvider struct {
 	error
 	ErrClass error
